+++ src/Contract/src/RoutesDelegator.php
@@ -26,6 +26,7 @@ use Frontend\Contract\Handler\Document\GetViewDocumentHandler;
 use Frontend\Contract\Handler\Document\PostCreateCommentHandler;
 use Frontend\Contract\Handler\Document\PostUploadFileHandler;
 use Frontend\Contract\Handler\Route\PostRouteContractHandler;
+use Frontend\Contract\Handler\Route\PostRouteSignCancelHandler;
 use Frontend\Contract\Handler\Route\PostRouteSignHandler;
 use Frontend\Contract\Handler\UCGS\GetCreateUCGSFormHandler;
 use Frontend\Contract\Handler\UCGS\PostCreateUCGSHandler;
@@ -33,6 +34,8 @@ use Frontend\Contract\Handler\UCLTS\GetCreateUCLTSFormHandler;
 use Frontend\Contract\Handler\UCLTS\PostCreateUCLTSHandler;
 use Frontend\Contract\Middleware\MetadataCorrectionMiddleware;
 use Frontend\Contract\Middleware\NotificationMiddleware;
+use Frontend\Contract\Middleware\PostRouteContractMiddleware;
+use Frontend\Contract\Middleware\PostRouteSignMiddleware;
 use Mezzio\Application;
 use Psr\Container\ContainerExceptionInterface;
 use Psr\Container\ContainerInterface;
@@ -87,7 +90,7 @@ class RoutesDelegator
             ->get('/list', GetListContractHandler::class, 'contract::list-contract')
             ->get('/view/' . $box_id, [MetadataCorrectionMiddleware::class, GetViewContractHandler::class], 'contract::view-contract-form');
 
-        $routeCollector->group('/route')->setMiddleware(NotificationMiddleware::class)
+        $routeCollector->group('/route')->setMiddleware([PostRouteContractMiddleware::class, NotificationMiddleware::class])
             ->get('/dept/' . $box_id, PostRouteContractHandler::class, 'route::dept')
             ->get('/legal/' . $box_id , PostRouteContractHandler::class, 'route::legal')
             ->get('/risk/' . $box_id , PostRouteContractHandler::class, 'route::risk')
@@ -96,8 +99,11 @@ class RoutesDelegator
             ->get('/vendor/' . $box_id , PostRouteContractHandler::class, 'route::vendor')
             ->get('/reject/' . $box_id , PostRouteContractHandler::class, 'route::reject');
         
-        $routeCollector->group('/sign')
-            ->post('/contract/{folder_id:[0-9-]*}/{file_id:[0-9-]*}', PostRouteSignHandler::class, 'route::sign');
+        $routeCollector->group('/sign')->setMiddleware([PostRouteSignMiddleware::class, NotificationMiddleware::class])
+            ->post('/contract/{folder_id:[0-9-]*}/{file_id:[0-9-]*}', PostRouteSignHandler::class, 'route::sign')
+        ;
+        $routeCollector
+            ->get('/cancel/sign/' . $box_id, PostRouteSignCancelHandler::class, 'route::sign-cancel');
         
         $routeCollector->group('/document')
             ->get('/edit/' . $box_id, GetEditDocumentHandler::class, 'document::edit-document')